Refactor further to make fizzbuzz a factory
