+++ index.html
@@ -20,7 +20,6 @@
         </form>
       </section>
     </main>
-    <!-- // <script src="js/fizzbuzz.js"></script> -->
     <script src="js/main.js"></script>
   </body>
 </html>
